+++ src/game.js
@@ -250,10 +250,6 @@ Game.prototype = {
   },
 
   isIllegalAt: function(y, x) {
-    if (this._moves.length === 0) {
-      return false;
-    }
-
     return this._ruleset.isIllegal(y, x, this);
   },