+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * When zoomed in with Tenuki's own zoom, pinch-zooming out will zoom the board out.
 * Simplify textured stone styling.
 * Add a new default renderer which uses SVG.
+* Support for komi as a `game.setup()` option.
 
 # v0.2.0