+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Next version
 
+* ...
+
+# v0.2.1
+
 * Resizing the window (including rotating) will re-calculate board sizing.
 * When zoomed in with Tenuki's own zoom, pinch-zooming out will zoom the board out.
 * Simplify textured stone styling.